Extra pieces added in
Block chance & dodge chance introduced based on Strength and Dexterity, also a spell has been included with a mana system.
next steps: work on a mana regen system for intelligence and also work on a luck system to increase the odds of block and dodge chances based on a luck stat.